requirements.txt
qrcode[pil]>=7.4.2
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
pyzbar>=0.1.9
tqdm>=4.65.0
click>=8.1.0
PyYAML>=6.0
opencv-contrib-python>=4.8.0
